These files need to be moved to resolve circular dependencies
