+++ hexrd/core/fitting/calibration/structureless.py
@@ -4,21 +4,10 @@ from typing import Optional
 import lmfit
 import numpy as np
 
-from hexrd.instrument import switch_xray_source
+from hexrd.core.instrument import switch_xray_source
 
-from .lmfit_param_handling import (
-    add_engineering_constraints,
-    create_instr_params,
-    create_tth_parameters,
-    DEFAULT_EULER_CONVENTION,
-    tth_parameter_prefixes,
-    update_instrument_from_params,
-)
-from .relative_constraints import (
-    create_relative_constraints,
-    RelativeConstraints,
-    RelativeConstraintsType,
-)
+from .lmfit_param_handling import add_engineering_constraints, create_instr_params, create_tth_parameters, DEFAULT_EULER_CONVENTION, tth_parameter_prefixes, update_instrument_from_params
+from ....core.fitting.calibration.relative_constraints import create_relative_constraints, RelativeConstraints, RelativeConstraintsType
 
 
 class StructurelessCalibrator:
